+++ app/Modules/Language/routes.php
@@ -1,8 +1,10 @@
 <?php
 
-use Illuminate\Support\Facades\Route;
 use App\Modules\Language\Controllers\LanguageController;
+use Illuminate\Support\Facades\Route;
 
 Route::middleware('web')->group(function () {
-    Route::get('/language/{lang?}', [LanguageController::class, 'index'])->name('language');
+    Route::get('/language/{language}', [LanguageController::class, 'index'])
+        ->whereIn('language', ['en', 'my'])
+        ->name('language');
 });